Fixed logic issue
